requirements.txt
pdfplumber==0.11.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy>=1.26.0
flask==3.0.3
